[REF] main컨트롤러 리팩토링
- Add LottoGame class
